fix(i18n): code-climate issues
